+++ execution/movr/update_location.cpp
@@ -6,7 +6,7 @@ namespace movr {
 
 UpdateLocationTxn::UpdateLocationTxn(const std::shared_ptr<StorageAdapter>& storage_adapter,
   const std::string& city, const uint64_t ride_id, const uint64_t timestamp,
-  const double lat, const double lon)
+  const uint64_t lat, const uint64_t lon)
     : location_histories_ (storage_adapter) {
   a_city_ = MakeFixedTextScalar<64>(city);
   a_ride_id_ = MakeInt64Scalar(ride_id);
